document_ai_project/requirements.txt
django
chromadb==0.4.15
openai
pytesseract
Pillow
python-dotenv
djangorestframework
pytest-django